requirements.txt
python-telegram-bot==20.7
APScheduler==3.10.4
python-dotenv==1.0.0
fastapi==0.109.2
uvicorn==0.27.1
sqlalchemy==2.0.27
jinja2==3.1.3
python-multipart==0.0.9
aiosqlite==0.20.0 